Move FetchUserService out of internal users folder
We're adding a new screen to allow users to see user notifications. We'll need to access the username of the related user, so we'll want to reuse the FetchUserService.

Our intent is that the new page can be used for both internal and external users, so we'll move the FetchUserService out of the internal folder to reflect this.
